feat: rename demoDocument to sample
